+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 _Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository._
 
+# Unreleased
+
+- Truncate long versions in `TaskGroupsTable` (\#1084).
+
 # 1.27.3
 
 - Add support for duplicating workflowtasks (\#1077).
 - Add `fractalJobId` for `/api/admin/v2/accounting/slurm` request (\#1081).
 - Add new users-info admin page (\#1079).
 - Fix query by email in admin's project page (\#1083).
-- Truncate long versions in `TaskGroupsTable` (\#1084).
 
 # 1.27.2