+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 _Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository._
 
+- Improve set/unset of JSON schema nullable properties (\#1085).
+- Show errors for unsupported JSON schema properties (\#1085).
+
 # 1.27.3
 
 - Add support for duplicating workflowtasks (\#1077).