+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 _Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository._
 
-# Unreleased
+# 1.27.4
 
 - Truncate long versions in `TaskGroupsTable` (\#1084).
 - Support creating a Workflow Template from Workflow JSON file (\#1076).